requirements.txt
websocket-client==1.7.0
psycopg2-binary==2.9.9
numpy==1.26.4